requirements.txt
# Core Python Libraries
python-dotenv==1.0.0
typing-extensions==4.9.0

# Web Framework
fastapi==0.109.0
uvicorn==0.24.0

# Natural Language Processing
nltk==3.8.1
spacy==3.7.3
transformers==4.37.2

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2

# Data Manipulation
pandas==2.2.0
numpy==1.26.3

# Utility and Helper Libraries
pydantic==2.6.4
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2

# Logging and Debugging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Security
cryptography==42.0.2

# Optional: AI and LLM Integration
anthropic==0.7.0

# Development and Productivity
black==24.1.1
isort==5.13.2
flake8==7.0.0

# Additional Libraries
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
boto3==1.34.113
botocore==1.34.113
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
cohere==5.5.3
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.30.0
crewai-tools==0.2.6
dataclasses-json==0.6.6
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.15
docx2txt==0.8
embedchain==0.1.102
exceptiongroup==1.2.1
fastavro==1.9.4
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.52.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gptcache==0.1.43
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib-resources==6.4.0
iniconfig==2.0.0
instructor==0.5.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
lancedb==0.5.7
langchain==0.1.5
langchain-cohere==0.1.5
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.4
langchain-ollama==0.1.0
langchain-text-splitters==0.0.2
langsmith==0.1.63
load-dotenv==0.1.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
onnxruntime==1.18.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pluggy==1.5.0
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.5.0
py==1.11.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic_core==2.18.2
PyGithub==1.59.1
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.9.18
PyNaCl==1.5.0
pyparsing==3.1.2
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyright==1.1.364
pysbd==0.3.4
PySocks==1.7.1
pytest==8.2.1
ratelimiter==1.2.0.post0
regex==2023.12.25
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
schema==0.7.7
selenium==4.16.1
semver==3.0.2
sentence-transformers==2.5.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.12
tenacity==8.3.0
tiktoken==0.5.2
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typer==0.9.4
types-requests==2.32.0.20240523
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.2.1
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.19.0
crewai
requests
beautifulsoup4
pandas
nltk
streamlit 
langchain_ollama
selenium
beautifulsoup4
lxml 
html5lib
pytz==2024.1
ollama==0.1.7
langchain-google-genai==1.0.4
langchain-ollama==0.1.0
google-generativeai==0.5.4

# Environment and Configuration
python-dotenv==1.0.0

# AI and Workflow Libraries
crewai==0.30.0
streamlit==1.32.0

# Web Automation
selenium==4.16.1
webdriver-manager==4.0.1  # Optional, can be skipped

# Web Interface
streamlit==1.29.0
